Remove /blog path to reduce nesting
I moved all the files inside blog folder to the root of routes and
changed the file paths accordingly. This will remove a lot of nested
folders inside blog folder since there won't we a lot of routes
besides /blog in the root either. It will make the permalinks clean
too and easier to remember