Remove transition wrapper to fix pages not exported
After trying to revert to my previous commits, seeing that upgrading
my packages or changing the config would cause the problem. But it turns
out that the transition was the culprit.

All the links inside the transition component wrapper was not detected
by the exporter/crawler and after I removed the wrapper from the each
routes, the export worked again.

My temporary workaround now is to not use any transition for pages with
any links in it.