fix: alias typescript unexpected token error
by logic, it seems we need to resolve aliases asap
but that doesn't seem to be the case here...

after debugging for a couple of hours I figured it out
this issue pops up after adding and using alias plugin.
the error shows up on build, so it must be rollup config,
debugging tsconfig doesn't make sense anymore because it
only provides intellisense and no more than that

fiddling around with plugins order, I tried to put
alias config in every possible position and found out it
works after the typescript plugin. this is one of the
most confusing things, but what do I know. I'm just glad
that it works now and I won't have to deal with this again.
just for the sake of symmetry, I reordered it in client too.